requirements.txt
networkx~=2.8.4
matplotlib~=3.5.1
numpy~=1.23.1
ruamel.yaml